Add a new backend attribute in VCL: "connect_timeout".
This is how long time we wait for a TCP connection to the backend to
become established.

Typical usage:

		backend b1 {
			.host = "hex";
			.port = "80";
			.connect_timeout = 500 ms;
		}

It can also be used in backends in director declarations.

Also add a parameter called "connect_timeout" which sets the default
to 400 msec (a number pulled out of my old black magicians hat).




git-svn-id: http://varnish.projects.linpro.no/svn/trunk@2642 d4fa192b-c00b-0410-8231-f00ffab90ce4