Try quite hard to allocate all storage segments for an object from the same stevedore as the object was allocated from.
I am not sure why I didn't spot this back when we added multiple
storage ability, but it is surely dumb to have to find the object
structure on one disk and the body of the object on another, or
even several other, disks.

For persistence this changes from a performance issue to problem:
Any object spread over silos will be lost on restart.

Further changes to this stuff is therefore in the pipeline, but I
wanted this fix in a version that can be merged back to 2.0.x.





git-svn-id: http://varnish.projects.linpro.no/svn/trunk@3872 d4fa192b-c00b-0410-8231-f00ffab90ce4