Begin moving bereq into the worker thread.
The original expectation was a ratio of 1:10 bereq:worker but once again
reality lets down data-less speculation.

It might have been true in a scenario where the majority of objects
bind the worker thread for significant time during delivery (ie:
ISOs over 28.8 modems) but that is not what happens in reality.

With objects rapidly delivered, backend contacts will size the worker
thread pool, since that is where they spend their time "out of the loop",
and consequenty, ratios of bereq/worker is much close to 1:1 than 1:10.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@3874 d4fa192b-c00b-0410-8231-f00ffab90ce4