Remove the vcl_timeout{} VCL callback.
We had big plans for this originally, but none of them have materialized
because there are better ways to do those things.

For instance: why invent a lot of code to do prefetch, when wget(1) can
do the job already ?

Getting rid of vcl_timeout{} (and vcl_discard{}, see yesterdays commit)
simplifies the expiry codes locking and statekeeping significantly.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@4100 d4fa192b-c00b-0410-8231-f00ffab90ce4